+++ packages/uui-base/lib/utils/index.ts
@@ -2,3 +2,4 @@ export * from './Timer';
 export * from './demandCustomElement';
 export * from './drag';
 export * from './math';
+export * from './findAncestorByAttributeValue';