+++ packages/uui-input/lib/uui-input.element.ts
@@ -1,6 +1,6 @@
 import { FormControlMixin, LabelMixin } from '@umbraco-ui/uui-base/lib/mixins';
 import { defineElement } from '@umbraco-ui/uui-base/lib/registration';
-import { css, html, LitElement, nothing, PropertyValueMap } from 'lit';
+import { css, html, LitElement, PropertyValueMap } from 'lit';
 import { property, query } from 'lit/decorators.js';
 import { ifDefined } from 'lit/directives/if-defined.js';
 
@@ -58,6 +58,8 @@ export class UUIInputElement extends FormControlMixin(
           var(--uui-input-border-color, var(--uui-color-border));
 
         --uui-button-height: 100%;
+        --auto-width-text-margin-right: 0;
+        --auto-width-text-margin-left: 0;
       }
 
       #control {
@@ -75,6 +77,8 @@ export class UUIInputElement extends FormControlMixin(
         z-index: -1;
         height: 0px;
         padding: 0 var(--uui-size-space-3);
+        margin: 0 var(--auto-width-text-margin-right) 0
+          var(--auto-width-text-margin-left);
       }
 
       :host([auto-width]) #input {
@@ -401,34 +405,41 @@ export class UUIInputElement extends FormControlMixin(
   render() {
     return html`
       ${this.renderPrepend()}
-      <div id="control">
-        <input
-          id="input"
-          .type=${this.type}
-          .value=${this.value as string}
-          .name=${this.name}
-          pattern=${ifDefined(this.pattern)}
-          min=${ifDefined(this.min)}
-          max=${ifDefined(this.max)}
-          step=${ifDefined(this.step)}
-          spellcheck=${this.spellcheck}
-          autocomplete=${ifDefined(this.autocomplete as any)}
-          placeholder=${ifDefined(this.placeholder)}
-          aria-label=${ifDefined(this.label)}
-          inputmode=${ifDefined(this.inputMode)}
-          ?disabled=${this.disabled}
-          ?autofocus=${this.autofocus}
-          ?required=${this.required}
-          ?readonly=${this.readonly}
-          @input=${this.onInput}
-          @change=${this.onChange} />
-        ${this.autoWidth ? this.renderAutoWidth() : nothing}
-      </div>
+      ${this.autoWidth ? this.renderInputWithAutoWidth() : this.renderInput()}
       ${this.renderAppend()}
     `;
   }
 
-  private renderAutoWidth() {
+  private renderInputWithAutoWidth() {
+    return html`<div id="control">
+      ${this.renderInput()}${this.renderAutoWidthBackground()}
+    </div>`;
+  }
+
+  renderInput() {
+    return html`<input
+      id="input"
+      .type=${this.type}
+      .value=${this.value as string}
+      .name=${this.name}
+      pattern=${ifDefined(this.pattern)}
+      min=${ifDefined(this.min)}
+      max=${ifDefined(this.max)}
+      step=${ifDefined(this.step)}
+      spellcheck=${this.spellcheck}
+      autocomplete=${ifDefined(this.autocomplete as any)}
+      placeholder=${ifDefined(this.placeholder)}
+      aria-label=${ifDefined(this.label)}
+      inputmode=${ifDefined(this.inputMode)}
+      ?disabled=${this.disabled}
+      ?autofocus=${this.autofocus}
+      ?required=${this.required}
+      ?readonly=${this.readonly}
+      @input=${this.onInput}
+      @change=${this.onChange} />`;
+  }
+
+  private renderAutoWidthBackground() {
     return html` <div id="auto" aria-hidden="true">${this.renderText()}</div>`;
   }